+++ test/widget/home/home.view_turno_test.dart
@@ -2,6 +2,7 @@ import 'package:appserap/database/app.database.dart';
 import 'package:appserap/database/respostas.database.dart';
 import 'package:appserap/dtos/prova.response.dto.dart';
 import 'package:appserap/dtos/prova_detalhes_caderno.response.dto.dart';
+import 'package:appserap/enums/download_status.enum.dart';
 import 'package:appserap/main.ioc.dart';
 import 'package:appserap/services/api.dart';
 import 'package:appserap/stores/home.store.dart';
@@ -22,17 +23,19 @@ import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:logging/logging.dart';
 
-import '../../../fixtures/fixture_reader.dart';
-import '../../../fixtures/user/http.fixture.dart';
-import '../../../fixtures/user/locator.fixture.dart';
-import 'home.view_test.mocks.dart';
+import '../../fixtures/fixture_reader.dart';
+import '../../fixtures/mocks.provas.dart';
+import '../../fixtures/mocks.provas.mocks.dart';
+import '../../fixtures/user/locator.fixture.dart';
+import 'home.view_turno_test.mocks.dart';
 
-@GenerateMocks([
-  ApiService,
-  PrincipalStore,
-  ProvaService,
-  QuestaoRespostaService,
-])
+@GenerateMocks(
+  [
+    ApiService,
+    PrincipalStore,
+    AppDatabase,
+  ],
+)
 void main() {
   setUpAll(() {
     initializeDateFormatting();
@@ -57,6 +60,7 @@ void main() {
 
     registerInjection<AppDatabase>(AppDatabase.executor(NativeDatabase.memory()));
     registerInjection<RespostasDatabase>(RespostasDatabase.executor(NativeDatabase.memory()));
+
     registerInjectionAsync<SharedPreferences>(() => SharedPreferences.getInstance());
 
     Future<UsuarioStore> teste() async {
@@ -83,65 +87,24 @@ void main() {
     unregisterInjection<RespostasDatabase>(disposingFunction: (p0) => p0.close());
   });
 
-  mockProvaSimples(MockProvaService mock) {
-    var json = fixture('prova_lista_simples.json');
-
-    List<ProvaResponseDTO> response = [];
-
-    for (var item in json) {
-      response.add(ProvaResponseDTO.fromJson(item));
-    }
-
-    when(mock.getProvas()).thenAnswer((realInvocation) async => buildResponse(response));
-  }
-
-  mockProvaNenhuma(MockProvaService mock) {
-    List<ProvaResponseDTO> response = [];
-
-    when(mock.getProvas()).thenAnswer((realInvocation) async => buildResponse(response));
-  }
-
   mockPrincipalStore() {
     when(ServiceLocator.get<PrincipalStore>().usuario).thenReturn(ServiceLocator.get<UsuarioStore>());
     when(ServiceLocator.get<PrincipalStore>().temConexao).thenReturn(true);
     when(ServiceLocator.get<PrincipalStore>().versao).thenReturn("");
   }
 
-  group('Home - Aba provas', () {
-    testWidgets('Não deve exibir nenuma prova na listagem', (tester) async {
+  group('Home - Turnos', () {
+    testWidgets('Deve exibir mensagem fora do horario da prova - Usuário turno tarde no horário da manhã',
+        (tester) async {
       await tester.runAsync(() async {
         await withClock(
-          Clock.fixed(DateTime(2022, 8, 31)),
+          Clock.fixed(DateTime(2022, 8, 31, 9, 0)),
           (() async {
             mockPrincipalStore();
 
             var provaMock = MockProvaService();
-            mockProvaNenhuma(provaMock);
-
-            when(ServiceLocator.get<ApiService>().prova).thenReturn(provaMock);
-
-            await tester.pumpWidget(MaterialApp(home: HomeView()));
-
-            await tester.pumpAndSettle(Duration(seconds: 20));
-          }),
-        );
-
-        expect(find.text('Prova atual'), findsOneWidget);
-        expect(find.byKey(Key('carregando')), findsNothing);
-        expect(find.byKey(Key('card-prova')), findsNWidgets(0));
-        expect(find.byKey(Key('sem-itens')), findsOneWidget);
-      });
-    });
-
-    testWidgets('Deve exibir uma prova na listagem', (tester) async {
-      await tester.runAsync(() async {
-        await withClock(
-          Clock.fixed(DateTime(2022, 8, 31)),
-          (() async {
-            mockPrincipalStore();
-
-            var provaMock = MockProvaService();
-            mockProvaSimples(provaMock);
+            await mockProvaComTempo(provaMock, forcarBaixada: true);
+            mockProvaDetalhesCaderno(provaMock);
 
             when(ServiceLocator.get<ApiService>().prova).thenReturn(provaMock);
 
@@ -155,18 +118,20 @@ void main() {
         expect(find.byKey(Key('carregando')), findsNothing);
         expect(find.byKey(Key('card-prova')), findsNWidgets(1));
         expect(find.byKey(Key('sem-itens')), findsNothing);
+        expect(find.byKey(Key('prova-indiponivel-turno')), findsOneWidget);
       });
     });
 
-    testWidgets('Não deve exibir nenuma prova antecipada na listagem', (tester) async {
+    testWidgets('Deve exibir prova para iniciar - Usuário turno tarde no horário da tarde', (tester) async {
       await tester.runAsync(() async {
         await withClock(
-          Clock.fixed(DateTime(2022, 8, 25)),
+          Clock.fixed(DateTime(2022, 8, 31, 15, 0)),
           (() async {
             mockPrincipalStore();
 
             var provaMock = MockProvaService();
-            mockProvaSimples(provaMock);
+            await mockProvaComTempo(provaMock, forcarBaixada: true);
+            mockProvaDetalhesCaderno(provaMock);
 
             when(ServiceLocator.get<ApiService>().prova).thenReturn(provaMock);
 
@@ -178,20 +143,24 @@ void main() {
 
         expect(find.text('Prova atual'), findsOneWidget);
         expect(find.byKey(Key('carregando')), findsNothing);
-        expect(find.byKey(Key('card-prova')), findsNWidgets(0));
-        expect(find.byKey(Key('sem-itens')), findsOneWidget);
+        expect(find.byKey(Key('card-prova')), findsNWidgets(1));
+        expect(find.byKey(Key('sem-itens')), findsNothing);
+        expect(find.byKey(Key('prova-indiponivel-turno')), findsNothing);
+        expect(find.byKey(Key('botao-prova-iniciar')), findsOneWidget);
       });
     });
 
-    testWidgets('Não deve exibir nenuma prova expirada na listagem', (tester) async {
+    testWidgets('Deve exibir mensagem fora do horario da prova - Usuário turno tarde no horário da noite',
+        (tester) async {
       await tester.runAsync(() async {
         await withClock(
-          Clock.fixed(DateTime(2022, 9, 1)),
+          Clock.fixed(DateTime(2022, 8, 31, 21, 0)),
           (() async {
             mockPrincipalStore();
 
             var provaMock = MockProvaService();
-            mockProvaSimples(provaMock);
+            await mockProvaComTempo(provaMock, forcarBaixada: true);
+            mockProvaDetalhesCaderno(provaMock);
 
             when(ServiceLocator.get<ApiService>().prova).thenReturn(provaMock);
 
@@ -203,33 +172,9 @@ void main() {
 
         expect(find.text('Prova atual'), findsOneWidget);
         expect(find.byKey(Key('carregando')), findsNothing);
-        expect(find.byKey(Key('card-prova')), findsNWidgets(0));
-        expect(find.byKey(Key('sem-itens')), findsOneWidget);
-      });
-    });
-
-    testWidgets('Não deve ser possivel iniciar a prova com tempo durante o final de semana', (tester) async {
-      await tester.runAsync(() async {
-        await withClock(
-          Clock.fixed(DateTime(2022, 9, 1)),
-          (() async {
-            mockPrincipalStore();
-
-            var provaMock = MockProvaService();
-            mockProvaSimples(provaMock);
-
-            when(ServiceLocator.get<ApiService>().prova).thenReturn(provaMock);
-
-            await tester.pumpWidget(MaterialApp(home: HomeView()));
-
-            await tester.pumpAndSettle(Duration(seconds: 20));
-          }),
-        );
-
-        expect(find.text('Prova atual'), findsOneWidget);
-        expect(find.byKey(Key('carregando')), findsNothing);
-        expect(find.byKey(Key('card-prova')), findsNWidgets(0));
-        expect(find.byKey(Key('sem-itens')), findsOneWidget);
+        expect(find.byKey(Key('card-prova')), findsNWidgets(1));
+        expect(find.byKey(Key('sem-itens')), findsNothing);
+        expect(find.byKey(Key('prova-indiponivel-turno')), findsOneWidget);
       });
     });
   });
