+++ test/widget/home/home.view_turno_test.mocks.dart
@@ -1,24 +1,20 @@
 // Mocks generated by Mockito 5.3.0 from annotations
-// in appserap/test/widget/login/home/home.view_test.dart.
+// in appserap/test/widget/home/home.view_turno_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i11;
+import 'dart:async' as _i12;
 
-import 'package:appserap/dtos/prova.response.dto.dart' as _i12;
-import 'package:appserap/dtos/prova_anterior.response.dto.dart' as _i16;
-import 'package:appserap/dtos/prova_detalhes.response.dto.dart' as _i13;
-import 'package:appserap/dtos/prova_detalhes_caderno.response.dto.dart' as _i14;
-import 'package:appserap/dtos/questao_resposta.dto.dart' as _i17;
-import 'package:appserap/dtos/questao_resposta.response.dto.dart' as _i15;
+import 'package:appserap/database/app.database.dart' as _i13;
 import 'package:appserap/services/api.dart' as _i3;
 import 'package:appserap/services/rest/admin.service.dart' as _i5;
 import 'package:appserap/services/rest/auth.admin.service.dart' as _i6;
 import 'package:appserap/services/rest/usuario.service.dart' as _i4;
-import 'package:appserap/stores/principal.store.dart' as _i9;
+import 'package:appserap/stores/principal.store.dart' as _i10;
 import 'package:appserap/stores/usuario.store.dart' as _i7;
 import 'package:chopper/chopper.dart' as _i2;
-import 'package:connectivity_plus/connectivity_plus.dart' as _i10;
+import 'package:connectivity_plus/connectivity_plus.dart' as _i11;
+import 'package:drift/drift.dart' as _i9;
 import 'package:mobx/mobx.dart' as _i8;
 import 'package:mockito/mockito.dart' as _i1;
 
@@ -141,14 +137,9 @@ class _FakeReactiveContext_18 extends _i1.SmartFake
       : super(parent, parentInvocation);
 }
 
-class _FakeType_19 extends _i1.SmartFake implements Type {
-  _FakeType_19(Object parent, Invocation parentInvocation)
-      : super(parent, parentInvocation);
-}
-
-class _FakeResponse_20<BodyType> extends _i1.SmartFake
-    implements _i2.Response<BodyType> {
-  _FakeResponse_20(Object parent, Invocation parentInvocation)
+class _FakeMigrationStrategy_19 extends _i1.SmartFake
+    implements _i9.MigrationStrategy {
+  _FakeMigrationStrategy_19(Object parent, Invocation parentInvocation)
       : super(parent, parentInvocation);
 }
 
@@ -249,7 +240,7 @@ class MockApiService extends _i1.Mock implements _i3.ApiService {
 /// A class which mocks [PrincipalStore].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockPrincipalStore extends _i1.Mock implements _i9.PrincipalStore {
+class MockPrincipalStore extends _i1.Mock implements _i10.PrincipalStore {
   MockPrincipalStore() {
     _i1.throwOnMissingStub(this);
   }
@@ -260,14 +251,14 @@ class MockPrincipalStore extends _i1.Mock implements _i9.PrincipalStore {
           returnValue: _FakeUsuarioStore_16(this, Invocation.getter(#usuario)))
       as _i7.UsuarioStore);
   @override
-  _i8.ObservableStream<_i10.ConnectivityResult> get conexaoStream =>
+  _i8.ObservableStream<_i11.ConnectivityResult> get conexaoStream =>
       (super.noSuchMethod(Invocation.getter(#conexaoStream),
-              returnValue: _FakeObservableStream_17<_i10.ConnectivityResult>(
+              returnValue: _FakeObservableStream_17<_i11.ConnectivityResult>(
                   this, Invocation.getter(#conexaoStream)))
-          as _i8.ObservableStream<_i10.ConnectivityResult>);
+          as _i8.ObservableStream<_i11.ConnectivityResult>);
   @override
   set conexaoStream(
-          _i8.ObservableStream<_i10.ConnectivityResult>? _conexaoStream) =>
+          _i8.ObservableStream<_i11.ConnectivityResult>? _conexaoStream) =>
       super.noSuchMethod(Invocation.setter(#conexaoStream, _conexaoStream),
           returnValueForMissingStub: null);
   @override
@@ -275,12 +266,12 @@ class MockPrincipalStore extends _i1.Mock implements _i9.PrincipalStore {
       super.noSuchMethod(Invocation.setter(#dispositivoId, _dispositivoId),
           returnValueForMissingStub: null);
   @override
-  _i10.ConnectivityResult get status =>
+  _i11.ConnectivityResult get status =>
       (super.noSuchMethod(Invocation.getter(#status),
-              returnValue: _i10.ConnectivityResult.bluetooth)
-          as _i10.ConnectivityResult);
+              returnValue: _i11.ConnectivityResult.bluetooth)
+          as _i11.ConnectivityResult);
   @override
-  set status(_i10.ConnectivityResult? _status) =>
+  set status(_i11.ConnectivityResult? _status) =>
       super.noSuchMethod(Invocation.setter(#status, _status),
           returnValueForMissingStub: null);
   @override
@@ -346,180 +337,55 @@ class MockPrincipalStore extends _i1.Mock implements _i9.PrincipalStore {
               returnValue: (Object? __p0, [Object? __p1, StackTrace? __p2]) {})
           as void Function(Object?, [Object?, StackTrace?]));
   @override
-  _i11.Future<void> setup() =>
+  _i12.Future<void> setup() =>
       (super.noSuchMethod(Invocation.method(#setup, []),
-              returnValue: _i11.Future<void>.value(),
-              returnValueForMissingStub: _i11.Future<void>.value())
-          as _i11.Future<void>);
+              returnValue: _i12.Future<void>.value(),
+              returnValueForMissingStub: _i12.Future<void>.value())
+          as _i12.Future<void>);
   @override
   void dispose() => super.noSuchMethod(Invocation.method(#dispose, []),
       returnValueForMissingStub: null);
   @override
-  _i11.Future<dynamic> onChangeConexao(_i10.ConnectivityResult? resultado) =>
+  _i12.Future<dynamic> onChangeConexao(_i11.ConnectivityResult? resultado) =>
       (super.noSuchMethod(Invocation.method(#onChangeConexao, [resultado]),
-          returnValue: _i11.Future<dynamic>.value()) as _i11.Future<dynamic>);
+          returnValue: _i12.Future<dynamic>.value()) as _i12.Future<dynamic>);
   @override
-  _i11.Future<void> obterVersaoDoApp() =>
+  _i12.Future<void> obterVersaoDoApp() =>
       (super.noSuchMethod(Invocation.method(#obterVersaoDoApp, []),
-              returnValue: _i11.Future<void>.value(),
-              returnValueForMissingStub: _i11.Future<void>.value())
-          as _i11.Future<void>);
+              returnValue: _i12.Future<void>.value(),
+              returnValueForMissingStub: _i12.Future<void>.value())
+          as _i12.Future<void>);
   @override
-  _i11.Future<void> sair() => (super.noSuchMethod(Invocation.method(#sair, []),
-          returnValue: _i11.Future<void>.value(),
-          returnValueForMissingStub: _i11.Future<void>.value())
-      as _i11.Future<void>);
+  _i12.Future<void> sair() => (super.noSuchMethod(Invocation.method(#sair, []),
+          returnValue: _i12.Future<void>.value(),
+          returnValueForMissingStub: _i12.Future<void>.value())
+      as _i12.Future<void>);
 }
 
-/// A class which mocks [ProvaService].
+/// A class which mocks [AppDatabase].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockProvaService extends _i1.Mock implements _i3.ProvaService {
-  MockProvaService() {
+class MockAppDatabase extends _i1.Mock implements _i13.AppDatabase {
+  MockAppDatabase() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i2.ChopperClient get client => (super.noSuchMethod(
-          Invocation.getter(#client),
-          returnValue: _FakeChopperClient_0(this, Invocation.getter(#client)))
-      as _i2.ChopperClient);
+  int get schemaVersion =>
+      (super.noSuchMethod(Invocation.getter(#schemaVersion), returnValue: 0)
+          as int);
   @override
-  set client(_i2.ChopperClient? _client) =>
-      super.noSuchMethod(Invocation.setter(#client, _client),
-          returnValueForMissingStub: null);
-  @override
-  Type get definitionType => (super.noSuchMethod(
-          Invocation.getter(#definitionType),
-          returnValue: _FakeType_19(this, Invocation.getter(#definitionType)))
-      as Type);
-  @override
-  _i11.Future<_i2.Response<List<_i12.ProvaResponseDTO>>> getProvas() => (super
-          .noSuchMethod(Invocation.method(#getProvas, []),
-              returnValue:
-                  _i11.Future<_i2.Response<List<_i12.ProvaResponseDTO>>>.value(
-                      _FakeResponse_20<List<_i12.ProvaResponseDTO>>(
-                          this, Invocation.method(#getProvas, []))))
-      as _i11.Future<_i2.Response<List<_i12.ProvaResponseDTO>>>);
-  @override
-  _i11.Future<_i2.Response<_i13.ProvaDetalhesResponseDTO>> getResumoProva(
-          {int? idProva}) =>
-      (super.noSuchMethod(Invocation.method(#getResumoProva, [], {#idProva: idProva}),
-              returnValue:
-                  _i11.Future<_i2.Response<_i13.ProvaDetalhesResponseDTO>>.value(
-                      _FakeResponse_20<_i13.ProvaDetalhesResponseDTO>(
-                          this,
-                          Invocation.method(
-                              #getResumoProva, [], {#idProva: idProva}))))
-          as _i11.Future<_i2.Response<_i13.ProvaDetalhesResponseDTO>>);
-  @override
-  _i11.Future<_i2.Response<_i14.ProvaDetalhesCadernoResponseDTO>> getResumoProvaCaderno(
-          {int? idProva, String? caderno}) =>
-      (super.noSuchMethod(Invocation.method(#getResumoProvaCaderno, [], {#idProva: idProva, #caderno: caderno}),
-              returnValue:
-                  _i11.Future<_i2.Response<_i14.ProvaDetalhesCadernoResponseDTO>>.value(
-                      _FakeResponse_20<_i14.ProvaDetalhesCadernoResponseDTO>(
-                          this,
-                          Invocation.method(
-                              #getResumoProvaCaderno, [], {#idProva: idProva, #caderno: caderno}))))
-          as _i11.Future<_i2.Response<_i14.ProvaDetalhesCadernoResponseDTO>>);
-  @override
-  _i11.Future<_i2.Response<int>> getStatusProva({int? idProva}) => (super
-      .noSuchMethod(Invocation.method(#getStatusProva, [], {#idProva: idProva}),
-          returnValue: _i11.Future<_i2.Response<int>>.value(
-              _FakeResponse_20<int>(
-                  this,
-                  Invocation.method(
-                      #getStatusProva, [], {#idProva: idProva})))) as _i11
-      .Future<_i2.Response<int>>);
-  @override
-  _i11.Future<_i2.Response<bool>> setStatusProva(
-          {int? idProva,
-          int? status,
-          int? tipoDispositivo,
-          int? dataInicio,
-          int? dataFim}) =>
-      (super.noSuchMethod(
-          Invocation.method(#setStatusProva, [], {
-            #idProva: idProva,
-            #status: status,
-            #tipoDispositivo: tipoDispositivo,
-            #dataInicio: dataInicio,
-            #dataFim: dataFim
-          }),
+  _i9.MigrationStrategy get migration => (super.noSuchMethod(
+          Invocation.getter(#migration),
           returnValue:
-              _i11.Future<_i2.Response<bool>>.value(_FakeResponse_20<bool>(
-                  this,
-                  Invocation.method(#setStatusProva, [], {
-                    #idProva: idProva,
-                    #status: status,
-                    #tipoDispositivo: tipoDispositivo,
-                    #dataInicio: dataInicio,
-                    #dataFim: dataFim
-                  })))) as _i11.Future<_i2.Response<bool>>);
+              _FakeMigrationStrategy_19(this, Invocation.getter(#migration)))
+      as _i9.MigrationStrategy);
   @override
-  _i11.Future<_i2.Response<List<_i15.QuestaoRespostaResponseDTO>>>
-      getRespostasPorProvaId({int? idProva}) => (super.noSuchMethod(
-          Invocation.method(#getRespostasPorProvaId, [], {#idProva: idProva}),
-          returnValue:
-              _i11.Future<_i2.Response<List<_i15.QuestaoRespostaResponseDTO>>>.value(
-                  _FakeResponse_20<List<_i15.QuestaoRespostaResponseDTO>>(this,
-                      Invocation.method(#getRespostasPorProvaId, [], {#idProva: idProva})))) as _i11
-          .Future<_i2.Response<List<_i15.QuestaoRespostaResponseDTO>>>);
+  _i12.Future<dynamic> limpar() =>
+      (super.noSuchMethod(Invocation.method(#limpar, []),
+          returnValue: _i12.Future<dynamic>.value()) as _i12.Future<dynamic>);
   @override
-  _i11.Future<_i2.Response<List<_i16.ProvaAnteriorResponseDTO>>>
-      getProvasAnteriores() => (super.noSuchMethod(
-          Invocation.method(#getProvasAnteriores, []),
-          returnValue:
-              _i11.Future<_i2.Response<List<_i16.ProvaAnteriorResponseDTO>>>.value(
-                  _FakeResponse_20<List<_i16.ProvaAnteriorResponseDTO>>(this,
-                      Invocation.method(#getProvasAnteriores, [])))) as _i11
-          .Future<_i2.Response<List<_i16.ProvaAnteriorResponseDTO>>>);
-}
-
-/// A class which mocks [QuestaoRespostaService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockQuestaoRespostaService extends _i1.Mock
-    implements _i3.QuestaoRespostaService {
-  MockQuestaoRespostaService() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i2.ChopperClient get client => (super.noSuchMethod(
-          Invocation.getter(#client),
-          returnValue: _FakeChopperClient_0(this, Invocation.getter(#client)))
-      as _i2.ChopperClient);
-  @override
-  set client(_i2.ChopperClient? _client) =>
-      super.noSuchMethod(Invocation.setter(#client, _client),
-          returnValueForMissingStub: null);
-  @override
-  Type get definitionType => (super.noSuchMethod(
-          Invocation.getter(#definitionType),
-          returnValue: _FakeType_19(this, Invocation.getter(#definitionType)))
-      as Type);
-  @override
-  _i11.Future<_i2.Response<_i15.QuestaoRespostaResponseDTO>> getRespostaPorQuestaoId(
-          {int? questaoId}) =>
-      (super.noSuchMethod(Invocation.method(#getRespostaPorQuestaoId, [], {#questaoId: questaoId}),
-              returnValue:
-                  _i11.Future<_i2.Response<_i15.QuestaoRespostaResponseDTO>>.value(
-                      _FakeResponse_20<_i15.QuestaoRespostaResponseDTO>(
-                          this,
-                          Invocation.method(
-                              #getRespostaPorQuestaoId, [], {#questaoId: questaoId}))))
-          as _i11.Future<_i2.Response<_i15.QuestaoRespostaResponseDTO>>);
-  @override
-  _i11.Future<_i2.Response<bool>> postResposta(
-          {String? chaveAPI, List<_i17.QuestaoRespostaDTO>? respostas}) =>
-      (super.noSuchMethod(Invocation.method(#postResposta, [], {#chaveAPI: chaveAPI, #respostas: respostas}),
-          returnValue: _i11.Future<_i2.Response<bool>>.value(
-              _FakeResponse_20<bool>(
-                  this,
-                  Invocation.method(#postResposta, [], {
-                    #chaveAPI: chaveAPI,
-                    #respostas: respostas
-                  })))) as _i11.Future<_i2.Response<bool>>);
+  _i12.Future<dynamic> limparBanco() =>
+      (super.noSuchMethod(Invocation.method(#limparBanco, []),
+          returnValue: _i12.Future<dynamic>.value()) as _i12.Future<dynamic>);
 }
